% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnp_function.R
\name{nnp}
\alias{nnp}
\title{A supervised projection function}
\usage{
nnp(x, y, hidden, steps)
}
\arguments{
\item{x}{is a data.frame that contains the inputs of the dataset}

\item{y}{is a vector that contains the outpus of the dataset}

\item{hidden}{is a vector that contains the number of neurons for each hidden layer}

\item{steps}{is the number of iterations of the net}
}
\description{
This function transforms a dataset into the activations of the neurons of the hidden layer of a neural network. This is done by training a neural network and then computing the activations of the neural network for each input pattern
}
\examples{
nnp_example()
}

