% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnp_function.R
\name{nnp2}
\alias{nnp2}
\title{A supervised projection function}
\usage{
nnp2(x, y, hidden, steps)
}
\arguments{
\item{x}{is a data.frame that contains the inputs of the dataset}

\item{y}{is a vector that contains the outpus of the dataset}

\item{hidden}{is a vector that contains the number of neurons for each hidden layer}

\item{rep}{is the number of iterations used for training the neural network}
}
\description{
This function transforms a dataset into the activations of the neurons of the hidden layer of a neural network. This is done by training a neural network and then computing the activations of the neural network for each input pattern
}
\examples{
data <- data.frame(A=runif(10, 5.0, 7.5), B=runif(10, 1.0, 3.5), C=runif(10, 10.0, 10.5), output=runif(10,1.0, 10.0))
projection <- nnp2(data[-ncol(data)], data[[ncol(data)]], hidden=2, rep=10)
}

